fixup! MAILBOX-314 search: precise spec about search on email and fix ES impl
