+++ mailbox/elasticsearch/src/main/java/org/apache/james/mailbox/elasticsearch/MailboxMappingFactoryV2.java
@@ -20,6 +20,8 @@
 package org.apache.james.mailbox.elasticsearch;
 
 import static org.apache.james.backends.es.IndexCreationFactory.CASE_INSENSITIVE;
+import static org.apache.james.backends.es.IndexCreationFactory.KEEP_MAIL_AND_URL;
+import static org.apache.james.backends.es.IndexCreationFactory.SNOWBALL_KEEP_MAIL_AND_URL;
 import static org.apache.james.backends.es.NodeMappingFactory.ANALYZER;
 import static org.apache.james.backends.es.NodeMappingFactory.BOOLEAN;
 import static org.apache.james.backends.es.NodeMappingFactory.FIELDS;
@@ -32,9 +34,8 @@ import static org.apache.james.backends.es.NodeMappingFactory.NOT_ANALYZED;
 import static org.apache.james.backends.es.NodeMappingFactory.PROPERTIES;
 import static org.apache.james.backends.es.NodeMappingFactory.RAW;
 import static org.apache.james.backends.es.NodeMappingFactory.SEARCH_ANALYZER;
-import static org.apache.james.backends.es.NodeMappingFactory.SPLIT_EMAIL;
 import static org.apache.james.backends.es.NodeMappingFactory.SNOWBALL;
-import static org.apache.james.backends.es.NodeMappingFactory.STANDARD;
+import static org.apache.james.backends.es.NodeMappingFactory.SPLIT_EMAIL;
 import static org.apache.james.backends.es.NodeMappingFactory.STRING;
 import static org.apache.james.backends.es.NodeMappingFactory.TYPE;
 import static org.apache.james.mailbox.elasticsearch.json.JsonMessageConstants.BCC;
@@ -192,163 +193,163 @@ public class MailboxMappingFactoryV2 implements MailboxMappingFactory {
                             .endObject()
 
                             .startObject(TO)
-                            .field(TYPE, NESTED)
-                            .startObject(PROPERTIES)
-                            .startObject(EMailer.NAME)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, KEEP_MAIL_AND_URL)
-                            .startObject(FIELDS)
-                            .startObject(RAW)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, CASE_INSENSITIVE)
-                            .endObject()
-                            .endObject()
-                            .endObject()
-                            .startObject(EMailer.ADDRESS)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, STANDARD)
-                            .field(NodeMappingFactory.SEARCH_ANALYZER, KEEP_MAIL_AND_URL)
-                            .startObject(FIELDS)
-                            .startObject(RAW)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, CASE_INSENSITIVE)
-                            .endObject()
-                            .endObject()
-                            .endObject()
-                            .endObject()
+                                .field(TYPE, NESTED)
+                                .startObject(PROPERTIES)
+                                    .startObject(EMailer.NAME)
+                                        .field(TYPE, STRING)
+                                        .field(ANALYZER, KEEP_MAIL_AND_URL)
+                                        .startObject(FIELDS)
+                                            .startObject(RAW)
+                                                .field(TYPE, STRING)
+                                                .field(ANALYZER, CASE_INSENSITIVE)
+                                            .endObject()
+                                        .endObject()
+                                    .endObject()
+                                    .startObject(EMailer.ADDRESS)
+                                        .field(TYPE, STRING)
+                                        .field(ANALYZER, STANDARD)
+                                        .field(NodeMappingFactory.SEARCH_ANALYZER, KEEP_MAIL_AND_URL)
+                                        .startObject(FIELDS)
+                                            .startObject(RAW)
+                                                .field(TYPE, STRING)
+                                                .field(ANALYZER, CASE_INSENSITIVE)
+                                            .endObject()
+                                        .endObject()
+                                    .endObject()
+                                .endObject()
                             .endObject()
 
                             .startObject(CC)
-                            .field(TYPE, NESTED)
-                            .startObject(PROPERTIES)
-                            .startObject(EMailer.NAME)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, KEEP_MAIL_AND_URL)
-                            .startObject(FIELDS)
-                            .startObject(RAW)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, CASE_INSENSITIVE)
-                            .endObject()
-                            .endObject()
-                            .endObject()
-                            .startObject(EMailer.ADDRESS)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, STANDARD)
-                            .field(NodeMappingFactory.SEARCH_ANALYZER, KEEP_MAIL_AND_URL)
-                            .startObject(FIELDS)
-                            .startObject(RAW)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, CASE_INSENSITIVE)
-                            .endObject()
-                            .endObject()
-                            .endObject()
-                            .endObject()
+                                .field(TYPE, NESTED)
+                                .startObject(PROPERTIES)
+                                    .startObject(EMailer.NAME)
+                                        .field(TYPE, STRING)
+                                        .field(ANALYZER, KEEP_MAIL_AND_URL)
+                                        .startObject(FIELDS)
+                                            .startObject(RAW)
+                                                .field(TYPE, STRING)
+                                                .field(ANALYZER, CASE_INSENSITIVE)
+                                            .endObject()
+                                        .endObject()
+                                    .endObject()
+                                    .startObject(EMailer.ADDRESS)
+                                        .field(TYPE, STRING)
+                                        .field(ANALYZER, STANDARD)
+                                        .field(NodeMappingFactory.SEARCH_ANALYZER, KEEP_MAIL_AND_URL)
+                                        .startObject(FIELDS)
+                                            .startObject(RAW)
+                                                .field(TYPE, STRING)
+                                                .field(ANALYZER, CASE_INSENSITIVE)
+                                            .endObject()
+                                        .endObject()
+                                    .endObject()
+                                .endObject()
                             .endObject()
 
                             .startObject(BCC)
-                            .field(TYPE, NESTED)
-                            .startObject(PROPERTIES)
-                            .startObject(EMailer.NAME)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, KEEP_MAIL_AND_URL)
-                            .startObject(FIELDS)
-                            .startObject(RAW)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, CASE_INSENSITIVE)
-                            .endObject()
-                            .endObject()
-                            .endObject()
-                            .startObject(EMailer.ADDRESS)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, STANDARD)
-                            .field(SEARCH_ANALYZER, KEEP_MAIL_AND_URL)
-                            .startObject(FIELDS)
-                            .startObject(RAW)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, CASE_INSENSITIVE)
-                            .endObject()
-                            .endObject()
-                            .endObject()
-                            .endObject()
+                                .field(TYPE, NESTED)
+                                    .startObject(PROPERTIES)
+                                    .startObject(EMailer.NAME)
+                                        .field(TYPE, STRING)
+                                        .field(ANALYZER, KEEP_MAIL_AND_URL)
+                                        .startObject(FIELDS)
+                                            .startObject(RAW)
+                                                .field(TYPE, STRING)
+                                                .field(ANALYZER, CASE_INSENSITIVE)
+                                            .endObject()
+                                        .endObject()
+                                    .endObject()
+                                    .startObject(EMailer.ADDRESS)
+                                        .field(TYPE, STRING)
+                                        .field(ANALYZER, STANDARD)
+                                        .field(SEARCH_ANALYZER, KEEP_MAIL_AND_URL)
+                                        .startObject(FIELDS)
+                                            .startObject(RAW)
+                                                .field(TYPE, STRING)
+                                                .field(ANALYZER, CASE_INSENSITIVE)
+                                            .endObject()
+                                        .endObject()
+                                    .endObject()
+                                .endObject()
                             .endObject()
 
                             .startObject(MAILBOX_ID)
-                            .field(TYPE, STRING)
-                            .field(INDEX, NOT_ANALYZED)
+                                .field(TYPE, STRING)
+                                .field(INDEX, NOT_ANALYZED)
                             .endObject()
 
                             .startObject(USERS)
-                            .field(TYPE, STRING)
-                            .field(INDEX, NOT_ANALYZED)
+                                .field(TYPE, STRING)
+                                .field(INDEX, NOT_ANALYZED)
                             .endObject()
 
                             .startObject(PROPERTIES)
-                            .field(TYPE, NESTED)
-                            .startObject(PROPERTIES)
-                            .startObject(Property.NAMESPACE)
-                            .field(TYPE, STRING)
-                            .field(INDEX, NOT_ANALYZED)
-                            .endObject()
-                            .startObject(Property.NAME)
-                            .field(TYPE, STRING)
-                            .field(INDEX, NOT_ANALYZED)
-                            .endObject()
-                            .startObject(Property.VALUE)
-                            .field(TYPE, STRING)
-                            .field(INDEX, NOT_ANALYZED)
-                            .endObject()
-                            .endObject()
+                                .field(TYPE, NESTED)
+                                .startObject(PROPERTIES)
+                                    .startObject(Property.NAMESPACE)
+                                        .field(TYPE, STRING)
+                                        .field(INDEX, NOT_ANALYZED)
+                                    .endObject()
+                                    .startObject(Property.NAME)
+                                        .field(TYPE, STRING)
+                                        .field(INDEX, NOT_ANALYZED)
+                                    .endObject()
+                                    .startObject(Property.VALUE)
+                                        .field(TYPE, STRING)
+                                        .field(INDEX, NOT_ANALYZED)
+                                    .endObject()
+                                .endObject()
                             .endObject()
 
                             .startObject(TEXT_BODY)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, KEEP_MAIL_AND_URL)
-                            .startObject(FIELDS)
-                            .startObject(NodeMappingFactory.SPLIT_EMAIL)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, STANDARD)
-                            .field(SEARCH_ANALYZER, KEEP_MAIL_AND_URL)
-                            .endObject()
-                            .startObject(RAW)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, CASE_INSENSITIVE)
-                            .field(IGNORE_ABOVE, MAXIMUM_TERM_LENGTH)
-                            .endObject()
-                            .endObject()
+                                .field(TYPE, STRING)
+                                .field(ANALYZER, KEEP_MAIL_AND_URL)
+                                .startObject(FIELDS)
+                                    .startObject(NodeMappingFactory.SPLIT_EMAIL)
+                                        .field(TYPE, STRING)
+                                        .field(ANALYZER, STANDARD)
+                                        .field(SEARCH_ANALYZER, KEEP_MAIL_AND_URL)
+                                    .endObject()
+                                    .startObject(RAW)
+                                        .field(TYPE, STRING)
+                                        .field(ANALYZER, CASE_INSENSITIVE)
+                                        .field(IGNORE_ABOVE, MAXIMUM_TERM_LENGTH)
+                                    .endObject()
+                                .endObject()
                             .endObject()
 
                             .startObject(HTML_BODY)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, KEEP_MAIL_AND_URL)
-                            .startObject(FIELDS)
-                            .startObject(SPLIT_EMAIL)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, STANDARD)
-                            .field(SEARCH_ANALYZER, KEEP_MAIL_AND_URL)
-                            .endObject()
-                            .startObject(RAW)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, CASE_INSENSITIVE)
-                            .field(IGNORE_ABOVE, MAXIMUM_TERM_LENGTH)
-                            .endObject()
-                            .endObject()
+                                .field(TYPE, STRING)
+                                .field(ANALYZER, KEEP_MAIL_AND_URL)
+                                .startObject(FIELDS)
+                                    .startObject(SPLIT_EMAIL)
+                                        .field(TYPE, STRING)
+                                        .field(ANALYZER, STANDARD)
+                                        .field(SEARCH_ANALYZER, KEEP_MAIL_AND_URL)
+                                    .endObject()
+                                    .startObject(RAW)
+                                        .field(TYPE, STRING)
+                                        .field(ANALYZER, CASE_INSENSITIVE)
+                                        .field(IGNORE_ABOVE, MAXIMUM_TERM_LENGTH)
+                                    .endObject()
+                                .endObject()
                             .endObject()
 
                             .startObject(HAS_ATTACHMENT)
-                            .field(TYPE, BOOLEAN)
+                                .field(TYPE, BOOLEAN)
                             .endObject()
 
                             .startObject(TEXT)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, SNOWBALL_KEEP_MAIL_AND_URL)
-                            .field(IGNORE_ABOVE, MAXIMUM_TERM_LENGTH)
-                            .startObject(FIELDS)
-                            .startObject(SPLIT_EMAIL)
-                            .field(TYPE, STRING)
-                            .field(ANALYZER, SNOWBALL)
-                            .field(SEARCH_ANALYZER, SNOWBALL_KEEP_MAIL_AND_URL)
-                            .endObject()
-                            .endObject()
+                                .field(TYPE, STRING)
+                                .field(ANALYZER, SNOWBALL_KEEP_MAIL_AND_URL)
+                                .field(IGNORE_ABOVE, MAXIMUM_TERM_LENGTH)
+                                .startObject(FIELDS)
+                                    .startObject(SPLIT_EMAIL)
+                                        .field(TYPE, STRING)
+                                        .field(ANALYZER, SNOWBALL)
+                                        .field(SEARCH_ANALYZER, SNOWBALL_KEEP_MAIL_AND_URL)
+                                    .endObject()
+                                .endObject()
                             .endObject()
                         .endObject()
                     .endObject()
