+++ mailbox/elasticsearch/src/main/java/org/apache/james/mailbox/elasticsearch/events/ElasticSearchListeningMessageSearchIndex.java
@@ -76,7 +76,7 @@ public class ElasticSearchListeningMessageSearchIndex extends ListeningMessageSe
             SearchCapabilities.FullText,
             SearchCapabilities.PartialEmailMatch);
 
-        if (messageToElasticSearchJson.handleIndexAttachment()) {
+        if (messageIndexer.supportsAttachments()) {
             supportedCapabilites.add(SearchCapabilities.Attachment);
         }
