+++ mailbox/elasticsearch/src/test/java/org/apache/james/mailbox/elasticsearch/ElasticSearchV2IntegrationTestWithoutTika.java
@@ -19,7 +19,6 @@
 
 package org.apache.james.mailbox.elasticsearch;
 
-import java.net.URISyntaxException;
 import java.time.ZoneId;
 
 import org.apache.james.mailbox.elasticsearch.json.MessageToElasticSearchJson;
@@ -31,40 +30,21 @@ import org.apache.james.mailbox.elasticsearch.search.ElasticSearchSearcherV1;
 import org.apache.james.mailbox.extractor.TextExtractor;
 import org.apache.james.mailbox.model.MailboxId;
 import org.apache.james.mailbox.model.MessageId;
-import org.apache.james.mailbox.tika.TikaConfiguration;
-import org.apache.james.mailbox.tika.TikaContainer;
-import org.apache.james.mailbox.tika.TikaHttpClientImpl;
-import org.apache.james.mailbox.tika.TikaTextExtractor;
+import org.apache.james.mailbox.store.extractor.DefaultTextExtractor;
 import org.elasticsearch.client.Client;
-import org.junit.Assume;
-import org.junit.ClassRule;
-import org.junit.Ignore;
-import org.junit.Test;
-import org.testcontainers.shaded.com.google.common.base.Throwables;
 
-public class ElasticSearchV1IntegrationTestWithTika extends  ElasticSearchIntegrationTest {
+public class ElasticSearchV2IntegrationTestWithoutTika extends  ElasticSearchIntegrationTest {
 
     private static final int SEARCH_SIZE = 1;
 
-    @ClassRule
-    public static TikaContainer tika = new TikaContainer();
-
     @Override
     protected TextExtractor getTextExtractor() {
-        try {
-            return new TikaTextExtractor(new TikaHttpClientImpl(TikaConfiguration.builder()
-                .host(tika.getIp())
-                .port(tika.getPort())
-                .timeoutInMillis(tika.getTimeoutInMillis())
-                .build()));
-        } catch (URISyntaxException e) {
-            throw Throwables.propagate(e);
-        }
+        return new DefaultTextExtractor();
     }
 
     @Override
     protected MailboxMappingFactory provideMappingFactory() {
-        return new MailboxMappingFactoryV1();
+        return new MailboxMappingFactoryV2();
     }
 
     @Override
@@ -80,35 +60,7 @@ public class ElasticSearchV1IntegrationTestWithTika extends  ElasticSearchIntegr
     @Override
     protected MessageToElasticSearchJson provideMessageToElasticSearchJson(TextExtractor textExtractor,
                                                                            ZoneId zoneId) {
-        return new MessageToElasticSearchJsonV1(textExtractor, zoneId, IndexAttachments.YES);
+        return new MessageToElasticSearchJsonV1(textExtractor, zoneId, IndexAttachments.NO);
     }
 
-
-    @Override
-    @Test
-    @Ignore("See MAILBOX-314 and upgrade to ElasticSearch schema version 2")
-    public void sortOnFromShouldWork() throws Exception {
-        Assume.assumeTrue(false);
-    }
-
-    @Override
-    @Test
-    @Ignore("See MAILBOX-314 and upgrade to ElasticSearch schema version 2")
-    public void searchWithTextShouldReturnMailsWhenCcMatches() throws Exception {
-        Assume.assumeTrue(false);
-    }
-
-    @Override
-    @Test
-    @Ignore("See MAILBOX-314 and upgrade to ElasticSearch schema version 2")
-    public void addressShouldReturnUidHavingRightExpeditorWhenFromIsSpecifiedWithDomainPartOfEmail() throws Exception {
-        Assume.assumeTrue(false);
-    }
-
-    @Override
-    @Test
-    @Ignore("See MAILBOX-314 and upgrade to ElasticSearch schema version 2")
-    public void searchShouldBeExactOnEmail() {
-        Assume.assumeTrue(false);
-    }
 }
